requirements.txt
django
djangorestframework
django-rest-knox
django-debug-toolbar
django-rest-swagger
